+++ src/h5xxhsum/cli/__init__.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2024-present S. Miccoli <stefano.miccoli@polimi.it>
 #
 # SPDX-License-Identifier: MIT
+import json
 import re
 import sys
 
@@ -16,8 +17,9 @@ digestre = re.compile(r"([0-9a-f]{32})\s+(.+)\n")
 @click.argument("h5", nargs=-1)
 @click.option("--check", "-c", multiple=False, type=click.File("r"))
 @click.option("--chunked/--no-chunked", default=False)
+@click.option("--jsonl/--no-jsonl", default=False)
 @click.version_option(version=__version__, prog_name="h5xxhsum")
-def h5xxhsum(h5, check, chunked):  # noqa: C901, PLR0912
+def h5xxhsum(h5, check, chunked, jsonl):  # noqa: C901, PLR0912
     if h5 and check:
         click.secho(
             "Cannot both verify and compute checksums",
@@ -39,7 +41,7 @@ def h5xxhsum(h5, check, chunked):  # noqa: C901, PLR0912
             ref, pth = mo.groups()
             to_be_verified += 1
             try:
-                msg = data_hash(pth, chunked)
+                msg, _ = data_hash(pth, chunked)
             except FileNotFoundError:
                 click.secho(f"{pth}: Missing", bold=True)
                 continue
@@ -55,11 +57,18 @@ def h5xxhsum(h5, check, chunked):  # noqa: C901, PLR0912
     else:
         for pth in h5:
             try:
-                msg = data_hash(pth, chunked)
+                msg, detail = data_hash(pth, chunked)
             except FileNotFoundError:
                 click.secho(f"{pth}: not found", file=sys.stderr, fg="red")
             except OSError as err:
                 click.secho(f"{pth}: {err}", file=sys.stderr, fg="red")
+
+            if jsonl:
+                click.echo(
+                    json.dumps(
+                        {"level0": {str(pth): msg}, "level1": dict(detail)}
+                    )
+                )
             else:
                 click.echo(f"{msg}  {pth}")
         sys.exit(0)
@@ -69,4 +78,4 @@ def data_hash(pth, chunked):
     callback = Walker(chunked=chunked)
     with h5py.File(pth, "r") as h5:
         h5.visititems(callback)
-    return callback.hexdigest
+    return callback.hexdigest, callback.level1